[win10] filesystem: move implementation to platform folder
