+++ xbmc/platform/win10/filesystem/WinLibraryDirectory.h
