+++ xbmc/platform/win10/filesystem/WinLibraryFile.cpp
