+++ xbmc/storage/win10/Win10StorageProvider.cpp
@@ -20,8 +20,8 @@
 #include "Win10StorageProvider.h"
 #include "guilib/LocalizeStrings.h"
 #include "filesystem/SpecialProtocol.h"
-#include "filesystem/win10/WinLibraryDirectory.h"
 #include "platform/win10/AsyncHelpers.h"
+#include "platform/win10/filesystem/WinLibraryDirectory.h"
 #include "platform/win32/CharsetConverter.h"
 #include "storage/MediaManager.h"
 #include "utils/JobManager.h"
